backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sqlalchemy-cockroachdb
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic
python-multipart
cloudinary
requests
opencv-python
pandas
numpy
